works

+++ src/main/java/my/demo/blockchain_demo/service/core/contract/functions/ProposeCurrencyFunction.java
@@ -6,14 +6,15 @@ import org.web3j.abi.datatypes.Type;
 import org.web3j.abi.datatypes.Uint;
 import org.web3j.abi.datatypes.generated.Bytes32;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @AllArgsConstructor
 public class ProposeCurrencyFunction implements DeFiFunction{
     private String currency;
     private String contractAddress;
-    private long min;
-    private long max;
+    private BigInteger min;
+    private BigInteger max;
 
     // proposeCurrency(bytes32 _name, address _contractAddr, uint _minAmount, uint _maxAmount) external onlyTreasurer
     @Override
